Move EventBusHeaders to impl package
